+++ app/src/main/java/com/sopt/now/compose/ui/ch_password/ChPasswordScreen.kt
@@ -34,7 +34,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
 import com.sopt.now.compose.R
-import com.sopt.now.compose.data.network.request.RequestChPwDto
+import com.sopt.now.compose.network.request.RequestChPwDto
 import com.sopt.now.compose.util.noRippleClickable
 
 @Composable
