+++ app/src/main/java/com/sopt/now/compose/ui/login/LoginViewModel.kt
@@ -4,9 +4,9 @@ import android.util.Log
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.sopt.now.compose.data.UserRepository
-import com.sopt.now.compose.data.network.ServicePool
-import com.sopt.now.compose.data.network.reponse.ResponseDto
-import com.sopt.now.compose.data.network.request.RequestLoginDto
+import com.sopt.now.compose.network.service.ServicePool
+import com.sopt.now.compose.network.reponse.ResponseDto
+import com.sopt.now.compose.network.request.RequestLoginDto
 import com.sopt.now.compose.ui.AuthState
 import org.json.JSONObject
 import retrofit2.Call
