+++ app/src/main/java/com/sopt/now/compose/ui/signup/SignupViewModel.kt
@@ -5,9 +5,9 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.sopt.now.compose.data.User
 import com.sopt.now.compose.data.UserRepository
-import com.sopt.now.compose.data.network.request.RequestSignUpDto
-import com.sopt.now.compose.data.network.reponse.ResponseDto
-import com.sopt.now.compose.data.network.ServicePool
+import com.sopt.now.compose.network.request.RequestSignUpDto
+import com.sopt.now.compose.network.reponse.ResponseDto
+import com.sopt.now.compose.network.service.ServicePool
 import com.sopt.now.compose.ui.AuthState
 import org.json.JSONObject
 import retrofit2.Call
